+++ src/module/feature/loops/enums/website-name.enum.ts
@@ -10,4 +10,8 @@ export const WebsiteNameEnum = {
      * Represents the website "The Armored Patrol".
      */
     THE_ARMORED_PATROL: 'The Armored Patrol',
+    /**
+     * Represents the website "World Of Tanks".
+     */
+    WORLD_OF_TANKS: 'World Of Tanks',
 } as const;